Move scim package to frontend
Only imported in frontend, to simplify the dependency graph I'm moving it into the cmd/frontend/internal/auth package.

Test plan:

Just code moved, CI will find any issues.
