+++ cmd/frontend/internal/auth/scim/filter/op_binary.go
