+++ cmd/frontend/internal/auth/scim/filter/op_boolean.go
