+++ cmd/frontend/internal/auth/scim/filter/op_decimal.go
