+++ cmd/frontend/internal/auth/scim/filter/op_string.go
