+++ cmd/frontend/internal/auth/scim/filter/operators.go
