+++ cmd/frontend/internal/scim/BUILD.bazel
@@ -7,10 +7,10 @@ go_library(
     visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//cmd/frontend/enterprise",
+        "//cmd/frontend/internal/auth/scim",
         "//internal/codeintel",
         "//internal/conf/conftypes",
         "//internal/database",
         "//internal/observation",
-        "//internal/scim",
     ],
 )
